unify cDNA and gDNA processing
A possible optimization may be to only count UMIs if there is at least
one UMI block present in the config, this is TODO
